refactor(Sections): Sections 일급컬렉션 활용
